Improve the performance of parser

+++ snap-parse/src/main/java/org/snapscript/parse/Token.java
@@ -3,5 +3,5 @@ package org.snapscript.parse;
 public interface Token<T> {
    T getValue();
    Line getLine();
-   int getType();
+   short getType();
 }
